app/requirements.txt
clickhouse-driver
fastapi
fastavro
motor
numpy
pandas
protobuf
pyarrow
pydantic
uvicorn
loguru
python-multipart
multipart
psycopg2
psycopg2-binary
aiochclient
aiohttp
kafka-python
sqlparse
requests
pytest
asyncio
yaml

